requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
bleach==6.1.0
breaching==0.1.2
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.3
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
grpcio==1.65.4
grpcio-tools==1.65.4
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.17
kiwisolver==1.4.5
lazy_loader==0.4
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-protobuf==3.6.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
pypickle==1.1.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch-msssim==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
setuptools==72.2.0
sewar==0.4.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
termcolor==2.4.0
text-unidecode==1.3
tifffile==2024.7.24
tikzplotlib==0.10.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-protobuf==5.27.0.20240626
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
